+++ examples/authn/invite_actions.go
@@ -10,6 +10,7 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/option"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/pangea"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/service/authn"
 )
@@ -33,10 +34,11 @@ func main() {
 	}
 
 	// Create config and client
-	client := authn.New(&pangea.Config{
-		Token:  token,
-		Domain: os.Getenv("PANGEA_DOMAIN"),
-	})
+	config, err := pangea.NewConfig(option.WithToken(token), option.WithDomain(os.Getenv("PANGEA_DOMAIN")))
+	if err != nil {
+		log.Fatalf("expected no error got: %v", err)
+	}
+	client := authn.New(config)
 
 	// Requests examples...
 	fmt.Println("Invite user 1...")