+++ examples/authn/user_actions.go
@@ -10,6 +10,7 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/option"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/pangea"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/service/authn"
 )
@@ -162,10 +163,11 @@ func main() {
 	}
 
 	// Create config and client
-	client := authn.New(&pangea.Config{
-		Token:  token,
-		Domain: os.Getenv("PANGEA_DOMAIN"),
-	})
+	config, err := pangea.NewConfig(option.WithToken(token), option.WithDomain(os.Getenv("PANGEA_DOMAIN")))
+	if err != nil {
+		log.Fatalf("expected no error got: %v", err)
+	}
+	client := authn.New(config)
 
 	// Requests examples...
 	// User create...
@@ -180,7 +182,7 @@ func main() {
 		OldPassword: PASSWORD_OLD,
 		NewPassword: PASSWORD_NEW,
 	}
-	_, err := client.Client.Password.Change(ctx, input3)
+	_, err = client.Client.Password.Change(ctx, input3)
 	if err != nil {
 		fmt.Println("Something went wrong...")
 		fmt.Println(err)