+++ examples/authz/authz_cycle.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"time"
 
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/option"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/pangea"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/service/authz"
 )
@@ -37,20 +38,20 @@ func main() {
 	defer cancelFn()
 
 	// create a new AuthZ client with pangea token and domain
-	client := authz.New(&pangea.Config{
-		Token:              token,
-		Domain:             os.Getenv("PANGEA_DOMAIN"),
-		QueuedRetryEnabled: true,
-		PollResultTimeout:  120 * time.Second,
-		Retry:              true,
-		RetryConfig: &pangea.RetryConfig{
-			RetryMax: 4,
-		},
-	})
+	config, err := pangea.NewConfig(
+		option.WithToken(token),
+		option.WithDomain(os.Getenv("PANGEA_DOMAIN")),
+		option.WithQueuedRetryEnabled(true),
+		option.WithPollResultTimeout(120*time.Second),
+	)
+	if err != nil {
+		log.Fatalf("expected no error got: %v", err)
+	}
+	client := authz.New(config)
 
 	// Create tuples
 	fmt.Println("Creating tuples...")
-	_, err := client.TupleCreate(ctx, &authz.TupleCreateRequest{
+	_, err = client.TupleCreate(ctx, &authz.TupleCreateRequest{
 		Tuples: []authz.Tuple{
 			{
 				Resource: authz.Resource{