+++ examples/embargo/ip_check.go
@@ -7,6 +7,7 @@ import (
 	"log"
 	"os"
 
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/option"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/pangea"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/service/embargo"
 )
@@ -17,10 +18,11 @@ func main() {
 		log.Fatal("Unauthorized: No token present")
 	}
 
-	embargocli := embargo.New(&pangea.Config{
-		Token:  token,
-		Domain: os.Getenv("PANGEA_DOMAIN"),
-	})
+	config, err := pangea.NewConfig(option.WithToken(token), option.WithDomain(os.Getenv("PANGEA_DOMAIN")))
+	if err != nil {
+		log.Fatalf("expected no error got: %v", err)
+	}
+	embargocli := embargo.New(config)
 
 	ctx := context.Background()
 	input := &embargo.IPCheckRequest{