+++ examples/intel/defang/defang.go
@@ -9,6 +9,7 @@ import (
 	"os"
 	"strings"
 
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/option"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/pangea"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/service/domain_intel"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/service/url_intel"
@@ -46,10 +47,14 @@ func main() {
 		log.Fatal("Unauthorized: No token present")
 	}
 
-	urlc := url_intel.New(&pangea.Config{
-		Token:  token,
-		Domain: os.Getenv("PANGEA_DOMAIN"),
-	})
+	config, err := pangea.NewConfig(
+		option.WithToken(token),
+		option.WithDomain(os.Getenv("PANGEA_DOMAIN")),
+	)
+	if err != nil {
+		log.Fatalf("expected no error got: %v", err)
+	}
+	urlc := url_intel.New(config)
 
 	url := "http://113.235.101.11:54384"
 
@@ -80,10 +85,15 @@ func main() {
 		log.Fatal(err)
 	}
 	fmt.Println(domain)
-	domainc := domain_intel.New(&pangea.Config{
-		Token:  token,
-		Domain: os.Getenv("PANGEA_DOMAIN"),
-	})
+
+	config, err = pangea.NewConfig(
+		option.WithToken(token),
+		option.WithDomain(os.Getenv("PANGEA_DOMAIN")),
+	)
+	if err != nil {
+		log.Fatalf("expected no error got: %v", err)
+	}
+	domainc := domain_intel.New(config)
 
 	ctx = context.Background()
 	domainReq := &domain_intel.DomainReputationRequest{