+++ examples/prompt_guard/guard/main.go
@@ -6,6 +6,7 @@ import (
 	"log"
 	"os"
 
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/option"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/pangea"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/service/prompt_guard"
 )
@@ -18,10 +19,14 @@ func main() {
 		log.Fatal("missing Prompt Guard API token")
 	}
 
-	promptGuardClient := prompt_guard.New(&pangea.Config{
-		Token:  token,
-		Domain: os.Getenv("PANGEA_DOMAIN"),
-	})
+	config, err := pangea.NewConfig(
+		option.WithToken(token),
+		option.WithDomain(os.Getenv("PANGEA_DOMAIN")),
+	)
+	if err != nil {
+		log.Fatalf("expected no error got: %v", err)
+	}
+	promptGuardClient := prompt_guard.New(config)
 
 	input := &prompt_guard.GuardRequest{
 		Messages: []prompt_guard.Message{