+++ examples/vault/structured_encrypt.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"time"
 
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/option"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/pangea"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/service/vault"
 )
@@ -22,10 +23,14 @@ func main() {
 		log.Fatal("missing PANGEA_DOMAIN environment variable")
 	}
 
-	vaultClient := vault.New(&pangea.Config{
-		Token:  token,
-		Domain: domain,
-	})
+	config, err := pangea.NewConfig(
+		option.WithToken(token),
+		option.WithDomain(domain),
+	)
+	if err != nil {
+		log.Fatalf("expected no error got: %v", err)
+	}
+	vaultClient := vault.New(config)
 
 	ctx := context.Background()